Add, test and fix integration with grand-challenge procedure
