+++ pyproject.toml
@@ -9,6 +9,7 @@ license = {text = "Apache-2.0"}
 requires-python = ">=3.10, <4.0"
 dependencies = [
     "pydicom>=2.4.0",
+    "grand-challenge-dicom-de-id-procedure",
 ]
 readme = "README.md"
